requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
cmake==3.28.3
decorator==4.4.2
dlib==19.24.2
exceptiongroup==1.2.0
facedetector-py==0.0.4
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
gast==0.4.0
gdown==5.1.0
google-auth==2.28.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
iniconfig==2.0.0
joblib==1.3.2
keras==2.13.1
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
moviepy==1.0.3
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
proglog==0.1.10
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PySocks==1.7.1
pytest==8.0.1
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.14
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2023.7.10
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.5.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
